requirements.txt
django
djangorestframework
drf-spectacular
django-redis
django-leaflet
gdal